added HTTP status code & hot fix db query

+++ warehouseManagement/backend/routers/category.py
@@ -1,13 +1,9 @@
-from fastapi import Depends, FastAPI, HTTPException, status, APIRouter
+from fastapi import Depends, APIRouter, HTTPException
 from fastapi.encoders import jsonable_encoder
-from fastapi.middleware.cors import CORSMiddleware
-from datetime import timedelta
 from typing import Annotated
-from fastapi.security import OAuth2PasswordRequestForm
-from table_model import category_model, token_model, login_model
+from table_model import category_model, login_model
 from authorization import auth
 from database import database
-from database_connect import get_db
 
 router = APIRouter(prefix="/category")
 
@@ -15,10 +11,12 @@ db = database.Database(url="sqlite:///./database/warehouseManagment.db")
 __BLOCK = Annotated[login_model.LoginModel.Table, Depends(auth.getCurrentActiveUser)]
 
 @router.get("/", status_code=200)
-async def getCategory(block: __BLOCK, db: database.Database = Depends(get_db)):
+async def getCategory(block: __BLOCK):
     db.connect()
     response = db.select("Category")
     db.close()
+    if response is None or not response:
+        raise HTTPException(status_code=404, detail="Category not found")
     return response
 
 
@@ -27,6 +25,8 @@ async def getCategory(id_category: int, block: __BLOCK):
     db.connect()
     response = db.select("Category", {"id": id_category})
     db.close()
+    if response is None or not response:
+        raise HTTPException(status_code=404, detail="Category not found")
     return response
 
 
@@ -34,8 +34,13 @@ async def getCategory(id_category: int, block: __BLOCK):
 async def createCategory(data: category_model.CategoryModel().Create, block: __BLOCK):
     body = jsonable_encoder(data)
     db.connect()
+    check = db.select("Category", {"name": body["name"]})
+    if check:
+        raise HTTPException(status_code=400, detail="Category existed")
     response = db.create("Category", body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
 
 
@@ -46,8 +51,13 @@ async def updateCategory(data: category_model.CategoryModel().Update, block: __B
         if value is None:
             del body[key]
     db.connect()
+    check = db.select("Category", {"id": body["id"]})
+    if not check:
+        raise HTTPException(status_code=400, detail="Category not existed")
     response = db.update("Category", {"id": body["id"]}, body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
 
 
@@ -55,7 +65,12 @@ async def updateCategory(data: category_model.CategoryModel().Update, block: __B
 async def deleteCategory(data: category_model.CategoryModel().Delete, block: __BLOCK):
     body = jsonable_encoder(data)
     db.connect()
+    check = db.select("Category", {"id": body["id"]})
+    if check:
+        raise HTTPException(status_code=400, detail="Category not existed")
     response = db.delete("Category", body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
