+++ warehouseManagement/backend/routers/worker.py
@@ -1,13 +1,9 @@
-from fastapi import Depends, FastAPI, HTTPException, status, APIRouter
+from fastapi import Depends, APIRouter, HTTPException
 from fastapi.encoders import jsonable_encoder
-from fastapi.middleware.cors import CORSMiddleware
-from datetime import timedelta
 from typing import Annotated
-from fastapi.security import OAuth2PasswordRequestForm
-from table_model import worker_model, token_model, login_model
+from table_model import worker_model, login_model
 from authorization import auth
 from database import database
-from database_connect import get_db
 
 router = APIRouter(prefix="/worker")
 
@@ -16,10 +12,12 @@ __BLOCK = Annotated[login_model.LoginModel.Table, Depends(auth.getCurrentActiveU
 
 
 @router.get("/", status_code=200)
-async def getWorker(block: __BLOCK, db: database.Database = Depends(get_db)):
+async def getWorker(block: __BLOCK):
     db.connect()
     response = db.select("Workers")
     db.close()
+    if response is None or not response:
+        raise HTTPException(status_code=404, detail="Workers not found")
     return response
 
 
@@ -28,6 +26,8 @@ async def getWorker(id_worker: int, block: __BLOCK):
     db.connect()
     response = db.select("Workers", {"id": id_worker})
     db.close()
+    if response is None or not response:
+        raise HTTPException(status_code=404, detail="Workers not found")
     return response
 
 
@@ -35,8 +35,13 @@ async def getWorker(id_worker: int, block: __BLOCK):
 async def createWorker(data: worker_model.WorkerModel().Create, block: __BLOCK):
     body = jsonable_encoder(data)
     db.connect()
+    check = db.select("Workers", {"name": body["name"], "surname": body["surname"]})
+    if check:
+        raise HTTPException(status_code=400, detail="Workers existed")
     response = db.create("Workers", body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
 
 
@@ -47,8 +52,13 @@ async def updateWorker(data: worker_model.WorkerModel().Update, block: __BLOCK):
         if value is None:
             del body[key]
     db.connect()
+    check = db.select("Workers", {"id": body["id"]})
+    if not check:
+        raise HTTPException(status_code=400, detail="Workers not existed")
     response = db.update("Workers", {"id": body["id"]}, body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
 
 
@@ -56,7 +66,12 @@ async def updateWorker(data: worker_model.WorkerModel().Update, block: __BLOCK):
 async def deleteWorker(data: worker_model.WorkerModel().Delete, block: __BLOCK):
     body = jsonable_encoder(data)
     db.connect()
+    check = db.select("Workers", {"id": body["id"]})
+    if not check:
+        raise HTTPException(status_code=400, detail="Workers not existed")
     response = db.delete("Workers", body)
     db.close()
+    if response is None:
+        raise HTTPException(status_code=418, detail="I’m a teapot")
     return response
